+++ src/components/Home/Hero/StatsBar/index.tsx
